REFACTOR 	CODE DUPLICATES 	-extracted method giveEvaluationToCriterion() from RemoveCriterionSteps and RemoveEvaluationSteps and put it in CommonTestDataAndOperations 	-extracted method evaluate() from calcMedia() in Student.groovy 	and doMedia() in EvaluationsByCriterion and added it to StudentController
